requirements.txt
numpy>=1.21.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=2.12.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910
pydantic>=1.8.2
asyncio>=3.4.3
typing-extensions>=4.0.0
PyYAML>=6.0
fastapi>=0.68.0
uvicorn>=0.15.0
prometheus-client>=0.11.0
kubernetes>=18.20.0
pre-commit>=2.15.0
bandit>=1.7.0
safety>=1.10.3
locust>=2.8.3
hypothesis>=6.0.0
pytest-hypothesis>=0.6.0
pytest-asyncio>=0.18.0
pytest-cov>=2.12.0
